requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
configparser==3.5.0       # via flake8
coverage==4.2             # via pytest-cov
docutils==0.12
enum34==1.1.6             # via flake8
flake8==3.0.4
isort==4.2.5
mccabe==0.5.2             # via flake8
multilint==1.0.2
py==1.4.31                # via pytest
pycodestyle==2.0.0        # via flake8
pyflakes==1.2.3           # via flake8
Pygments==2.1.3
pytest-cov==2.3.1
pytest-django==3.0.0
pytest==3.0.2
sqlparse==0.2.1
